refactor(blocks-construct): revamp the dashboard to remove duplicate loc
