+++ src/features/dashboard/index.ts
@@ -1,1 +1,5 @@
 export { CircularProgress } from './component/circular-progress/circular-progress';
+export { DashboardOverview } from './component/dashboard-overview/dashboard-overview';
+export { DashboardUserPlatform } from './component/dashboard-user-platform/dashboard-user-platform';
+export { DashboardUserActivityGraph } from './component/dashboard-user-activity-graph/dashboard-user-activity-graph';
+export { DashboardSystemOverview } from './component/dashboard-system-overview/dashboard-system-overview';
